local_deployment/requirements.txt
notion-client
python-dotenv
aiohttp
watchdog
